Update and rename Task 3/Insertion Sort/insertion sort_c to Task 3/InsertionSort/Insertionsort.c
The output to the following input is- 5 6 11 12 13

Insertion sort :Insertion sort is a simple sorting algorithm that works similar to the way you sort playing cards in your hands. The array is virtually split into a sorted and an unsorted part. Values from the unsorted part are picked and placed at the correct position in the sorted part.